Pause Resume Works
I was finally able to get the coroutine problems to work that stemmed from GENERIC. the key was to use this. so that if you wanted to stop the method prematurely your new class that isn't generic started it or is the owner and thus can destroy it as well. such a great feeling